refactor: Major module reordering (get rid of separate application modulefor Application singleton, merge into main.rs of crate
