+++ src/game_renderers/diff_print_old.rs
@@ -19,9 +19,9 @@ use rand::RngExt;
 use super::*;
 
 use crate::{
-    application::TemporaryAppData,
     fmt_helpers::{fmt_button, fmt_button_ascii, fmt_duration, fmt_hertz, FmtTetromino},
     graphics_settings::Glyphset,
+    TemporaryAppData,
 };
 
 #[derive(
