+++ src/menus/adjust_gameplay.rs
@@ -18,15 +18,13 @@ use crossterm::{
 use falling_tetromino_engine::{ExtNonNegF64, RotationSystem, TetrominoGenerator};
 
 use crate::{
-    application::{
-        menus::{Menu, MenuUpdate},
-        Application, Settings,
-    },
     fmt_helpers::FmtBool,
+    menus::{Menu, MenuUpdate},
+    Application, Settings,
 };
 
 impl<T: Write> Application<T> {
-    pub(in crate::application) fn run_menu_adjust_gameplay(&mut self) -> io::Result<MenuUpdate> {
+    pub fn run_menu_adjust_gameplay(&mut self) -> io::Result<MenuUpdate> {
         let if_unmodifiable_clone_and_switch = |s: &mut Settings| {
             if let Some(cloned_slot_idx) = s
                 .gameplay_slotmachine
