+++ src/menus/advanced_settings.rs
@@ -13,16 +13,14 @@ use crossterm::{
 };
 
 use crate::{
-    application::{
-        menus::{Menu, MenuUpdate},
-        Application, SavefileGranularity,
-    },
     fmt_helpers::FmtBool,
     game_renderers::TetroTUIRenderer,
+    menus::{Menu, MenuUpdate},
+    Application, SavefileGranularity,
 };
 
 impl<T: Write> Application<T> {
-    pub(in crate::application) fn run_menu_advanced_settings(&mut self) -> io::Result<MenuUpdate> {
+    pub fn run_menu_advanced_settings(&mut self) -> io::Result<MenuUpdate> {
         let mut selected = 0usize;
         loop {
             let w_main = Self::W_MAIN.into();
